Erreur

+++ app/controller/jobController.js
@@ -21,81 +21,81 @@ console.log('ici');
     },
 
 
-
-    async addJobProject (req,res) {
-        const projectId= req.params.id;
-        const jobName = req.body.job
-        console.log('ici 2');
+    async addJobProject (req, res) {
         try {
+            const projectId= req.params.id;
+            if (!projectId){
+                throw new MainError('missing parameter', req, res, 400);
+            };
+            const jobName = req.body.job
             const jobfinded = await jobDatamapper.getJobId(jobName);
-            console.log(jobfinded.id);
+            if (!jobfinded){
+                throw new MainError('This job does not exists', req, res, 400);
+            };
             const result = await jobDatamapper.addJob (projectId ,jobfinded.id);
-            
-            if (result === null || result === undefined){
-                
+
+            if (!result){
+                throw new MainError('This job was not adding', req, res, 404);
+
             };
-
             return res.status(204).json(result);
-
         } catch (error) {
-            console.error(error);
-            return res.status(500).json({ message: "Database Error", error: error});
+        console.error(error);
         };
     },
 
-    async deleteJobProject (req,res) {
-        
-        const tableId = req.body.id_project_has_job;
-        
+    async deleteJobProject (req, res) {
         try {
-
-            
+            const tableId = req.body.id_project_has_job;
+            if (!tableId){
+                throw new MainError('missing parameter', req, res, 400);
+            };
             const result = await jobDatamapper.deleteJobProject (tableId);
-            if (result === null || result === undefined){
-                return res.status(404).json({ message: "This job does not exists !"});
+            if (!result){
+                throw new MainError('This job was not delete', req, res, 404);
             };
             return res.status(204).json(result);
-
         } catch (error) {
-            console.error(error);
-            return res.status(500).json({ message: "Database Error", error: error});
+        console.error(error);
         };
     },
         
 
-    async addJobUser (req,res) {
-        const userId= req.params.id;
-        const jobName = req.body.job
-        console.log('ici 2');
+    async addJobUser (req, res) {
         try {
+            const userId= req.params.id;
+            if (!userId){
+                throw new MainError('missing parameter', req, res, 400);
+            };
+            const jobName = req.body.job
+            console.log('ici 2');
             const jobfinded = await jobDatamapper.getJobId(jobName);
-            const result = await jobDatamapper.addJobUser (userId ,jobfinded.id);
-            if (result === null || result === undefined){
-                return res.status(404).json({ message: "This job does not exists !"});
+            if (!jobfinded){
+                throw new MainError('This job does not exists', req, res, 400);
             };
-
+            const result = await jobDatamapper.addJobUser (userId ,jobfinded.id);
+            if (!result){
+                throw new MainError('This job was not adding', req, res, 404);
+            };
             return res.status(204).json(result);;
-
         } catch (error) {
-            console.error(error);
-            return res.status(500).json({ message: "Database Error", error: error});
+        console.error(error);
         };
     },
 
-    async deleteJobUser (req,res) {
-        const userId= req.params.id;
-        
+    async deleteJobUser (req, res) {
         try {
-            const result = await jobDatamapper.deleteJobUser(userId);
-            if (result === null || result === undefined){
-                return res.status(404).json({ message: "This job does not exists !"});
+            const userId= req.params.id;
+            if (!userId){
+                throw new MainError('missing parameter', req, res, 400);
             };
-
+            const result = await jobDatamapper.deleteJobUser(userId);
+            if (!result){
+                throw new MainError('This job was not delete', req, res, 404);
+            };
             return res.status(204).json(result);
-
         } catch (error) {
             console.error(error);
-            return res.status(500).json({ message: "Database Error", error: error});
         };
     }
